server/requirements.txt
# MySQL MCP Server Requirements
# MCP(Model Context Protocol) 서버 실행에 필요한 패키지 목록

# MCP 관련 패키지
mcp>=1.0.0

# MySQL 연결
mysql-connector-python>=8.0.0

# OpenAI API (자연어 처리용)
openai>=1.0.0

# 비동기 처리
asyncio

# 로깅 및 유틸리티
typing-extensions>=4.0.0

# 개발 및 테스트용 (선택사항)
pytest>=7.0.0
pytest-asyncio>=0.21.0 